# upload/extracted/Attached-Assets-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Цифровой обходной лист — Digital Routing Sheet HR system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List candidates with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: branchId
          in: query
          schema:
            type: number
        - name: positionId
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CandidateWithSheet"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate card (recruiter)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get candidate by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Candidate with full routing sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update candidate info
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"

  # ── Offers ────────────────────────────────────────────
...
  /doctor-profiles/{routingSheetId}:
    get:
      operationId: getDoctorProfile
      tags: [doctor-profiles]
      summary: Get doctor profile for a routing sheet (chief physician writes, account manager reads)
      security:
        - bearerAuth: []
      parameters:
        - name: routingSheetId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertDoctorProfile
      tags: [doctor-profiles]
      summary: Create or update doctor profile (chief physician only)
      security:
        - bearerAuth: []
      parameters:
        - name: routingSheetId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorProfileInput"
      responses:
        "200":
          description: Doctor profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorProfile"

  # ── Termination Sheets ────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ─────────────────────────────────────────────
...
    Candidate:
      type: object
      required: [id, fullName, email, phone, offerStatus, createdAt]
      properties:
        id:
          type: number
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        experience:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        offerStatus:
          type: string
          enum: [draft, sent, accepted, expired]
        createdById:
          type: number
        createdAt:
          type: string
          format: date-time
    CandidateInput:
      type: object
      required: [fullName, email, phone]
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        experience:
          type: string
        education:
          type: string
        certifications:
          type: string
    CandidateUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        experience:
          type: string
        education:
          type: string
        certifications:
          type: string
    CandidateWithSheet:
      type: object
      required: [id, fullName, email, phone, offerStatus, createdAt]
      properties:
        id:
          type: number
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        offerStatus:
          type: string
          enum: [draft, sent, accepted, expired]
        routingSheet:
          $ref: "#/components/schemas/RoutingSheet"
        createdAt:
          type: string
          format: date-time
    CandidateDetail:
      type: object
      required: [id, fullName, email, phone, offerStatus, createdAt]
      properties:
        id:
          type: number
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        experience:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        offerStatus:
          type: string
          enum: [draft, sent, accepted, expired]
        createdById:
          type: number
        createdAt:
          type: string
          format: date-time
        routingSheet:
          $ref: "#/components/schemas/RoutingSheetDetail"

    # ── Offers ───────────────────────────────────────────
...
    RoutingSheet:
      type: object
      required: [id, candidateId, branchId, positionId, isDoctor, status, createdAt]
      properties:
        id:
          type: number
        candidateId:
          type: number
        candidateName:
          type: string
        branchId:
          type: number
        branchName:
          type: string
        positionId:
          type: number
        positionName:
          type: string
        isDoctor:
          type: boolean
        status:
          type: string
          enum: [in_progress, completed, cancelled]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    RoutingSheetDetail:
      type: object
      required: [id, candidateId, branchId, positionId, isDoctor, status, steps, createdAt]
      properties:
        id:
          type: number
        candidateId:
          type: number
        candidateName:
          type: string
        branchId:
          type: number
        branchName:
          type: string
        positionId:
          type: number
        positionName:
          type: string
        isDoctor:
          type: boolean
        status:
          type: string
          enum: [in_progress, completed, cancelled]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RoutingStep"
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    RoutingStep:
      type: object
      required: [id, routingSheetId, stepType, assignedRole, status, isBackground, createdAt]
      properties:
        id:
          type: number
        routingSheetId:
          type: number
        stepType:
          type: string
          enum: [hr_registration, marketing_photo, tb_briefing, it_accounts, audit_training, doctor_profile, site_publication, final_review]
        assignedRole:
          type: string
          enum: [admin, recruiter, hr, marketing, tb, it, audit, chief_physician, account_manager]
        assignedUserId:
          type: number
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        isBackground:
          type: boolean
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        completedById:
          type: number
        completedByName:
          type: ["string", "null"]
        isOverride:
          type: boolean
        stepData:
          type: ["object", "null"]
          additionalProperties: true
        completedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    DoctorProfile:
      type: object
      required: [id, routingSheetId, createdAt]
      properties:
        id:
          type: number
        routingSheetId:
          type: number
        experience:
          type: ["number", "null"]
        specialty:
          type: ["string", "null"]
        ageRestrictions:
          type: ["string", "null"]
        siteDiscounts:
          type: ["string", "null"]
        about:
          type: ["string", "null"]
        procedures:
          type: ["array", "null"]
          items:
            type: string
        photoUrl:
          type: ["string", "null"]
        createdById:
          type: ["number", "null"]
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    DoctorProfileInput:
      type: object
      properties:
        experience:
          type: number
        specialty:
          type: string
        ageRestrictions:
          type: string
        siteDiscounts:
          type: string
        about:
          type: string
        procedures:
          type: array
          items:
            type: string
        photoUrl:
          type: string

    # ── Termination Sheets ────────────────────────────────
...